support hadoop3, spark3.5, and GCS connector 3.0.0
The new GCS connector drops support for hadoop <3, changes the version
string, and, most importantly, substantially reduces the memory use of
trivial pipelines. I saw a reduction from >3.75GiB to ~1.3GiB on a
trivial Hail pipeline. I believe this is entirely due to better HTTP
buffer management in the GCS hadoop connector. Without this change,
in combination with Java 11 and Spark 3.5.0, we see a marginal but
meaningful increase in memory use. Enough to fail some of our somewhat
strict memory tests.